Update to photoz.tex
Minor updates to figure plot sizes (so that plots appear in a 3 column 2 row array and match the caption, not 2 column 3 row as they were previously rendering). Minor update to figure captions because the number of points in one set of the plots has changed (and now all runs match in terms of number of catalog galaxies used).