David's edits
--renu wrote:

"Since LSST will be its own follow-up machine with respect to detection and confirmation and orbit linkage for minor planets, I wondered if a brief summary of that effort by the project would be useful in this chapter."

i added a paragraph in S3.1 which i think might be okay, but i haven't
tried to make a "brief summary of that effort by the project". perhaps
this is something that lynne or zeljko could do (or we could just
skip it -- the rest of 3.1 does okay with describing MOPS etc.)

--references added:

Bolin et al. 2014 Icarus 241 280

--renu wrote:

"Some goals are listed in Table 3.1. It was unclear to me why these are the specific goals."

this is an item for either lynne or zeljko, or
else just leave it as is.

--renu wrote:

"Some simulation results are listed in Table 3.2. It would be helpful to give a brief statement on what was assumed about the simulated populations."

item for lynne

--renu wrote:

"In section 3.3.3, second paragraph was confusing."

agreed that this is confusing. item for lynne to address.

--renu wrote:

"Regarding “active asteroids”, I wondered if these could constitute a significant source of noise for extra-galactic science."

i would think not since active asteroids move and extragalactic
sources don't. not sure if/how to respond to this. thoughts?

--renu wrote:

"n section 3.6.3, first paragraph, “NEOs” should probably be “TNOs”."

well, it *should* be NEOs, i think, but the plot (figure 3.8)
shows no change for NEOs, but a big improvement for TNOs.
lynne, does this make sense to you? or a problem with the figure .. ?

--typos listed by renu have apparently been
fixed by meg schwamb.